feat(api,hook): add hook synchronization
