+++ engine/sql/api/246_project_repo_hook.sql
@@ -1,0 +1,5 @@
+-- +migrate Up
+ALTER TABLE project_repository ADD COLUMN hook_configuration JSONB;
+
+-- +migrate Down
+ALTER TABLE project_repository DROP COLUMN hook_configuration;
