+++ sdk/cdsclient/interface.go
@@ -310,6 +310,7 @@ type HookClient interface {
 	PollVCSEvents(uuid string, workflowID int64, vcsServer string, timestamp int64) (events sdk.RepositoryEvents, interval time.Duration, err error)
 	VCSConfiguration() (map[string]sdk.VCSConfiguration, error)
 	VCSGerritConfiguration() (map[string]sdk.VCSGerritConfiguration, error)
+	RepositoriesListAll(ctx context.Context) ([]sdk.ProjectRepository, error)
 }
 
 // ServiceClient exposes functions used for services
